Extract JwtService to own file
